move methods from UtilService to Factory
