+++ src/test/java/de/tum/in/www1/artemis/post/ConversationFactory.java
@@ -1,0 +1,40 @@
+package de.tum.in.www1.artemis.post;
+
+import java.time.ZoneId;
+import java.time.ZonedDateTime;
+import java.util.HashSet;
+import java.util.Set;
+
+import de.tum.in.www1.artemis.domain.User;
+import de.tum.in.www1.artemis.domain.enumeration.DisplayPriority;
+import de.tum.in.www1.artemis.domain.metis.Post;
+import de.tum.in.www1.artemis.domain.metis.Reaction;
+
+public class ConversationFactory {
+
+    private static int dayCount = 1;
+
+    public static Reaction createReactionForUser(User user) {
+        Reaction reaction = new Reaction();
+        reaction.setEmojiId("heart");
+        reaction.setUser(user);
+        return reaction;
+    }
+
+    public static Post createBasicPost(Integer i, User author) {
+        Post post = new Post();
+        post.setTitle(String.format("Title Post %s", (i + 1)));
+        post.setContent(String.format("Content Post %s", (i + 1)));
+        post.setVisibleForStudents(true);
+        post.setDisplayPriority(DisplayPriority.NONE);
+        post.setAuthor(author);
+        post.setCreationDate(ZonedDateTime.of(2015, 11, dayCount, 23, 45, 59, 1234, ZoneId.of("UTC")));
+        String tag = String.format("Tag %s", (i + 1));
+        Set<String> tags = new HashSet<>();
+        tags.add(tag);
+        post.setTags(tags);
+
+        dayCount = (dayCount % 25) + 1;
+        return post;
+    }
+}
